[hotfix] [state backends, tests] Make StateBackendMigrationTestBase more understandable
The StateBackendMigrationTestBase previously mocked user behaviour of
upgrading serializers by using a single serializer class, that can be
configured with different target compatibility results when they are
checked for compatibility.

This is a bit hard to understand, also doesn't really reflect how a user
would actually approach the feature. Instead, instead of configuring a
single serializer class with different compatibility "personalities",
this commit uses actual different classes, V1TestTypeSerializer,
V2TestTypeSerializer, and IncompatibleTestTypeSerializer, to simulate
the compatibility cases in tests.

This commit also refactors the serializer migration related test
serializers / types / snapshots to the testutil package, so that it can
be shared by other state migration related tests in the future.
